forgor to stagger

+++ Assets/Scripts/Objects/MeleeWeapon.cs
@@ -62,7 +62,7 @@ public class MeleeWeapon : Weapon
 		otherWeapon?.ActivateKnockBack();
 	}
 
-	protected override void OnAttack(int _Dmg)
+	protected override void OnAttack(float _Dmg)
 	{
 		firstHit = true;
 		otherWeapon?.OnAttack(_Dmg);
